test: dedup REQUIRED_ENV/_settings_env into conftest.py, drop redundant sys.path inserts
backend: unify _ingest_commits/_ingest_prs into one generic _ingest_items

Also fixes a pre-existing bug: test_github_client.py's fake_get helpers
didn't accept the timeout kwarg added in the previous commit, breaking
list_prs tests.
